+++ app/assets/javascripts/v1/application_common/old/controllers/ReceiveButtonCtrl.js.coffee
@@ -4,8 +4,8 @@ AnnictOld.angular.controller 'ReceiveButtonCtrl', ($scope, $http) ->
 
   $scope.toggle = (channelId) ->
     if $scope.isReceiving
-      $http.delete("/api/receptions/#{channelId}").success ->
+      $http.delete("/api/internal/receptions/#{channelId}").success ->
         $scope.isReceiving = false
     else
-      $http.post('/api/receptions', channel_id: channelId).success ->
+      $http.post('/api/internal/receptions', channel_id: channelId).success ->
         $scope.isReceiving = true
