+++ app/assets/javascripts/v1/application_common/old/directives/aniTipModal.js.coffee
@@ -1,10 +1,10 @@
-AnnictOld.angular.directive 'aniTipModal', ->
-  restrict: 'C'
+AnnictOld.angular.directive "aniTipModal", ->
+  restrict: "C"
   scope: true
 
   link: (scope, element, attributes) ->
     scope.display = true
-    scope.target = $(element).find('.tip').data('target')
+    scope.target = $(element).find(".tip").data("target")
 
   controller: ($scope, $element, $http) ->
     $scope.openModal = ->
@@ -12,6 +12,7 @@ AnnictOld.angular.directive 'aniTipModal', ->
       false
 
     $scope.finishTip = (partialName) ->
-      if confirm('非表示にします。よろしいですか？')
-        $http.post('/api/tips/finish', partial_name: partialName).success ->
+      if confirm("非表示にします。よろしいですか？")
+        path = "/api/internal/tips/finish"
+        $http.post(path, partial_name: partialName).success ->
           $scope.display = false
