+++ app/assets/javascripts/v1/application_common/old/directives/annChannelSelector.js.coffee
@@ -11,7 +11,7 @@ AnnictOld.angular.directive 'annChannelSelector', ->
       if $scope.prevChannelId != $scope.channelId
         $scope.$emit("showSpinner-#{workId}")
 
-        $http.post "/api/works/#{$scope.workId}/channels/select",
+        $http.post "/api/internal/works/#{$scope.workId}/channels/select",
           channel_id: $scope.channelId
         .success ->
           $scope.prevChannelId = $scope.channelId
