--- app/controllers/api/tips_controller.rb
@@ -1,9 +1,0 @@
-class Api::TipsController < ApplicationController
-  before_action :authenticate_user!, only: [:finish]
-
-  def finish(partial_name)
-    UserTipsService.new(current_user).finish!(partial_name)
-
-    render status: 200, nothing: true
-  end
-end
