+++ app/views/api/internal/latest_statuses/skip_episode.jbuilder
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+partial_path = "/api/internal/latest_statuses/latest_status"
+json.partial!(partial_path, latest_status: @latest_status)
