+++ app/views/api/internal/organizations/index.jbuilder
@@ -1,1 +1,3 @@
+# frozen_string_literal: true
+
 json.organizations @organizations, :id, :name
