+++ app/views/api/internal/people/index.jbuilder
@@ -1,1 +1,3 @@
+# frozen_string_literal: true
+
 json.people @people, :id, :name
