+++ app/views/api/internal/works/friends.jbuilder
