--- app/views/api/latest_statuses/skip_episode.jbuilder
@@ -1,1 +1,0 @@
-json.partial!("/api/latest_statuses/latest_status", latest_status: @latest_status)
