--- app/views/api/users/activities.jbuilder
@@ -1,1 +1,0 @@
-json.partial! '/api/activities/activities', activities: @activities
